Renames BayesianModel to BayesianNetwork
