+++ pgmpy/tests/test_estimators/test_ParameterEstimator.py
@@ -3,13 +3,13 @@ import unittest
 from pandas import DataFrame
 from numpy import NaN
 
-from pgmpy.models import BayesianModel
+from pgmpy.models import BayesianNetwork
 from pgmpy.estimators import ParameterEstimator
 
 
 class TestParameterEstimator(unittest.TestCase):
     def setUp(self):
-        self.m1 = BayesianModel([("A", "C"), ("B", "C"), ("D", "B")])
+        self.m1 = BayesianNetwork([("A", "C"), ("B", "C"), ("D", "B")])
         self.d1 = DataFrame(
             data={"A": [0, 0, 1], "B": [0, 1, 0], "C": [1, 1, 0], "D": ["X", "Y", "Z"]}
         )
