+++ pgmpy/tests/test_factors/test_discrete/test_Factor.py
@@ -11,7 +11,7 @@ from pgmpy.factors import factor_divide
 from pgmpy.factors import factor_product
 from pgmpy.factors.discrete.CPD import TabularCPD
 from pgmpy.independencies import Independencies
-from pgmpy.models import BayesianModel
+from pgmpy.models import BayesianNetwork
 from pgmpy.models import MarkovModel
 from pgmpy.utils import get_example_model
 
@@ -1590,7 +1590,7 @@ class TestJointProbabilityDistributionMethods(unittest.TestCase):
         )
 
     def test_is_imap(self):
-        G1 = BayesianModel([("diff", "grade"), ("intel", "grade")])
+        G1 = BayesianNetwork([("diff", "grade"), ("intel", "grade")])
         diff_cpd = TabularCPD("diff", 2, [[0.2], [0.8]])
         intel_cpd = TabularCPD("intel", 3, [[0.5], [0.3], [0.2]])
         grade_cpd = TabularCPD(
