+++ pgmpy/tests/test_utils/test_state_name.py
@@ -4,7 +4,7 @@ import numpy as np
 import numpy.testing as np_test
 
 from pgmpy.factors.discrete import DiscreteFactor, TabularCPD
-from pgmpy.models import BayesianModel
+from pgmpy.models import BayesianNetwork
 from pgmpy.inference import Inference
 from pgmpy.inference import VariableElimination
 
@@ -54,7 +54,7 @@ class TestStateNameInit(unittest.TestCase):
             state_names=self.sn2,
         )
 
-        student = BayesianModel([("diff", "grade"), ("intel", "grade")])
+        student = BayesianNetwork([("diff", "grade"), ("intel", "grade")])
         diff_cpd = TabularCPD("diff", 2, [[0.2], [0.8]])
         intel_cpd = TabularCPD("intel", 2, [[0.3], [0.7]])
         grade_cpd = TabularCPD(
@@ -119,8 +119,8 @@ class StateNameDecorator(unittest.TestCase):
             state_names=self.sn2,
         )
 
-        student = BayesianModel([("diff", "grade"), ("intel", "grade")])
-        student_state_names = BayesianModel([("diff", "grade"), ("intel", "grade")])
+        student = BayesianNetwork([("diff", "grade"), ("intel", "grade")])
+        student_state_names = BayesianNetwork([("diff", "grade"), ("intel", "grade")])
 
         diff_cpd = TabularCPD("diff", 2, [[0.2], [0.8]])
         intel_cpd = TabularCPD("intel", 2, [[0.3], [0.7]])
